requirements.txt
appdirs==1.4.4
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1673151334029/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1700874570898/work/dist/numpy-1.26.2-cp311-cp311-macosx_11_0_arm64.whl#sha256=6a1afc4f3b7776e43c5d2749d8fae541f47a49ecd1ec0fff58a7ea345a7a67b5
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
pydicom==2.4.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-ssim==0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
SciPy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1700812589137/work/dist/scipy-1.11.4-cp311-cp311-macosx_11_0_arm64.whl#sha256=7f8fc22aa24cb1ee5318d88f2162f00440d915ebd87d725a05a0fccec9246fb9
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
wandb==0.16.1
